Changed the way XL and IL is handled in CRAVA. New rules: 1. CRAVA will always read the grid geometry of the first seismic grid initially. 2. If this grid aligns with the inversion volume so that moving one cell in the i-index gives a constant XL or IL change, and moving one cell in the j index gives a constant change of the other, XL and IL information is taken from the geometry. Otherwise, the old XL = i, IL = j mapping is used.
The alignment check may be removed. CRAVA will then map each cell to closest XL/IL

git-svn-id: https://code.nr.no/scm/svn/crava/trunk@321 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/modelsettings.cpp
@@ -336,7 +336,7 @@ bool ModelSettings::noiseIsScaled()
 {
   if(noiseScaled_.size()>0)
   {
-    for(int i=0;i<noiseScaled_.size();i++)
+    for(int i=0;i<static_cast<int>(noiseScaled_.size());i++)
     {
       if(noiseScaled_[i]!=NULL)
         return true;
